+++ lib/compendium-api.rb
@@ -9,6 +9,7 @@ require 'nretnil/compendium-api/content_group'
 require 'nretnil/compendium-api/custom_field'
 require 'nretnil/compendium-api/export'
 require 'nretnil/compendium-api/publisher'
+require 'nretnil/compendium-api/role'
 require 'nretnil/compendium-api/user'
 require 'nretnil/compendium-api/version'
 # Helper Class